create a file for reducer
